# microservices/backend/openapi_server/openapi/openapi.yaml
openapi: 3.0.2
info:
  title: Main API
  version: 0.0.1
servers:
- url: https://reserva.emilharvey.ca
paths:
...
  /groups:
    get:
      operationId: groups_get
      parameters: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_groups_get_200_response'
          description: OK
      summary: Gets a list of Group objects
      tags:
      - group
      x-openapi-router-controller: openapi_server.controllers.group_controller
    post:
      operationId: groups_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewGroup'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_groups_post_201_response'
          description: OK
      tags:
      - group
      x-openapi-router-controller: openapi_server.controllers.group_controller
  /groups/{id}:
    delete:
      operationId: groups_id_delete
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          $ref: '#/components/schemas/GroupId'
        style: simple
      responses:
        "204":
          description: Deleted
      summary: Deletes a Group object
      tags:
      - group
      x-openapi-router-controller: openapi_server.controllers.group_controller
    get:
      operationId: groups_id_get
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          $ref: '#/components/schemas/GroupId'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_groups_post_201_response'
          description: OK
      summary: Gets a Group object by id
      tags:
      - group
      x-openapi-router-controller: openapi_server.controllers.group_controller
    put:
      operationId: groups_id_put
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          $ref: '#/components/schemas/GroupId'
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroup'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_groups_post_201_response'
          description: OK
      summary: Updates a Group object
      tags:
      - group
      x-openapi-router-controller: openapi_server.controllers.group_controller
  /groups/{id}/memberships:
    get:
      operationId: groups_id_memberships_get
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          $ref: '#/components/schemas/GroupId'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_groups__id__memberships_get_200_response'
          description: OK
      summary: Gets a list of GroupMembership objects
      tags:
      - group
      x-openapi-router-controller: openapi_server.controllers.group_controller
    post:
      operationId: groups_id_memberships_post
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          $ref: '#/components/schemas/GroupId'
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewGroupMembership'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_groups__id__memberships_post_201_response'
          description: OK
      summary: Creates a new GroupMembership object
      tags:
      - group
      x-openapi-router-controller: openapi_server.controllers.group_controller
  /groups/{id}/memberships/{membershipId}:
    delete:
      operationId: groups_id_memberships_membership_id_delete
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          $ref: '#/components/schemas/GroupId'
        style: simple
      - explode: false
        in: path
        name: membershipId
        required: true
        schema:
          $ref: '#/components/schemas/GroupMembershipId'
        style: simple
      responses:
        "204":
          description: Deleted
        "404":
          description: Not Found
      summary: Deletes a GroupMembership object
      tags:
      - group
      x-openapi-router-controller: openapi_server.controllers.group_controller
    get:
      operationId: groups_id_memberships_membership_id_get
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          $ref: '#/components/schemas/GroupId'
        style: simple
      - explode: false
        in: path
        name: membershipId
        required: true
        schema:
          $ref: '#/components/schemas/GroupMembershipId'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_groups__id__memberships_post_201_response'
          description: OK
        "404":
          description: Not Found
      summary: Gets a GroupMembership object by id
      tags:
      - group
      x-openapi-router-controller: openapi_server.controllers.group_controller
...
components:
  schemas:
...
    Group:
      allOf:
      - $ref: '#/components/schemas/NewGroup'
      - $ref: '#/components/schemas/Group_allOf'
      title: Group
    GroupId:
      description: |
        The ID of a group. This is a UUID with the prefix "group-".
      format: uuid
      pattern: "^[a-z]+-[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
      title: GroupId
      type: string
    GroupMembership:
      allOf:
      - $ref: '#/components/schemas/NewGroupMembership'
      - $ref: '#/components/schemas/GroupMembership_allOf'
      title: GroupMembership
    GroupMembershipId:
      description: The ID of a group membership.
      format: uuid
      pattern: "^[a-z]+-[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
      title: GroupMembershipId
      type: string
...
    NewGroup:
      allOf:
      - $ref: '#/components/schemas/UpdateGroup'
      description: |
        A group. This could be a project group, a team, or something else. This is sent by the client when creating a new group.
      required:
      - createdBy
      - name
      title: NewGroup
    NewGroupMembership:
      description: |
        This is a group membership. It is the intermediate object between a user and a group. It is used to determine which users are members of which groups.
      example:
        user: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        group: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        group:
          description: |
            The ID of a group. This is a UUID with the prefix "group-".
          format: uuid
          pattern: "^[a-z]+-[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          title: GroupId
          type: string
        user:
          description: |
            Id of a user. This is a UUID with the prefix "user-".
          format: uuid
          pattern: "^[a-z]+-[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          title: UserId
          type: string
      required:
      - group
      - user
      title: NewGroupMembership
      type: object
...
    UpdateGroup:
      description: |
        A group. This could be a project group, a team, or something else. This is sent by the client when updating an existing group.
      example:
        createdBy: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        name: Project Group
      properties:
        name:
          description: The name of the group.
          example: Project Group
          title: name
          type: string
        createdBy:
          description: |
            Id of a user. This is a UUID with the prefix "user-".
          format: uuid
          pattern: "^[a-z]+-[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          title: UserId
          type: string
      title: UpdateGroup
      type: object
...
    _groups_get_200_response:
      example:
        Groups:
        - null
        - null
      properties:
        Groups:
          items:
            $ref: '#/components/schemas/Group'
          title: Groups
          type: array
      title: _groups_get_200_response
    _groups_post_201_response:
      example:
        Group: null
      properties:
        Group:
          $ref: '#/components/schemas/Group'
      title: _groups_post_201_response
    _groups__id__memberships_get_200_response:
      example:
        GroupMemberships:
        - null
        - null
      properties:
        GroupMemberships:
          items:
            $ref: '#/components/schemas/GroupMembership'
          title: GroupMemberships
          type: array
      title: _groups__id__memberships_get_200_response
    _groups__id__memberships_post_201_response:
      example:
        GroupMembership: null
      properties:
        GroupMembership:
          $ref: '#/components/schemas/GroupMembership'
      title: _groups__id__memberships_post_201_response
...
    Group_allOf:
      description: |
        A group. This could be a project group, a team, or something else. The difference between this and NewGroup is that this has an ID, which is assigned by the server. It also has timestamps for when the group was created and last updated.
      properties:
        id:
          description: |
            The ID of a group. This is a UUID with the prefix "group-".
          format: uuid
          pattern: "^[a-z]+-[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          title: GroupId
          type: string
        createdAt:
          description: The date and time the group was created.
          example: 2019-01-01T00:00:00Z
          format: date-time
          readOnly: true
          title: createdAt
          type: string
        lastUpdatedAt:
          description: The date and time the group was last updated.
          example: 2019-03-01T00:00:00Z
          format: date-time
          readOnly: true
          title: lastUpdatedAt
          type: string
      required:
      - createdAt
      - id
      - lastUpdatedAt
      title: Group_allOf
      type: object
      example: null
    GroupMembership_allOf:
      properties:
        id:
          description: The ID of a group membership.
          format: uuid
          pattern: "^[a-z]+-[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          title: GroupMembershipId
          type: string
      title: GroupMembership_allOf
      type: object
      example: null
...